code clean up
changing the model
