+++ trainer.py
@@ -9,84 +9,10 @@ from model import SeLuModel
 from model import ReLuModel
 from model import Runner
 from constants import *
+from formatter import create_data
 
 
 DIRNAME = os.path.dirname(os.path.realpath(__file__))
-
-"""
-Formats the data into x_val and y_val
-@param JSON value
-@returns x_val, y_val
-"""
-def format(data):
-	home = []
-	away = []
-	# winner = [0] # Winner : 1 if home 0 if away
-	winner = [data['score_board']['summary']['home']['r']] # Home team's score
-
-	# Get the winner
-	# if data['score_board']['summary']['home']['r'] > data['score_board']['summary']['away']['r']:
-	# 	winner[0] = 1
-
-	for k1, v1 in data['score_board']['summary'].items():
-		if k1 == 'home':
-			for k2, v2 in v1.items():
-				if k2 != 'r':
-					home.append(v2)
-		else:
-			for k2, v2 in v1.items():
-				if k2 != 'r':
-					away.append(v2)
-
-	for k1, v1 in data['pitcher_info'].items():
-		if k1 == 'home':
-			for k2, v2 in v1[0].items():
-				if k2 != 'name':
-					if k2 == 'era':
-						home.append(float(v2))
-					else:	
-						home.append(v2)
-		else:
-			for k2, v2 in v1[0].items():
-				if k2 != 'name':
-					if k2 == 'era':
-						away.append(float(v2))
-					else:	
-						away.append(v2)
-
-	for k1, v1 in data['batter_info'].items():
-			if k1 == 'home':
-				for k2, v2 in v1[0].items():
-					if k2 != 'name':
-						if k2 == 'hra':
-							home.append(float(v2))
-						else:
-							home.append(v2)
-			else:
-				for k2, v2 in v1[0].items():
-					if k2 != 'name':
-						if k2 == 'hra':
-							away.append(float(v2))
-						else:
-							away.append(v2)
-
-	for k1, v1 in data['away_team_standing'].items():
-		if k1 != 'name':
-			away.append(v1)
-
-	for k1, v1 in data['home_team_standing'].items():
-		if k1 != 'name':
-			home.append(v1)
-
-	return home + away, winner
-
-
-# home, away, y = format(data[0])
-# print(len(home), len(away))
-# print(home)
-# print(away)
-# print(y)
-
 
 parser = argparse.ArgumentParser(description='KBO Score Prediction Trainer')
 
@@ -107,19 +33,7 @@ if __name__ == '__main__':
 	print("Load JSON data")
 	data = json.load(f)
 
-	dataX = []
-	dataY = []
-
-	for i in range(len(data)):
-		# Put home team data
-		x_val, y_val = format(data[i])
-		dataX.append(x_val)
-		dataY.append(y_val)
-
-	train_size = int(len(dataY) * args.train_size)
-	test_size = len(dataY) - train_size
-	trainX, trainY = np.array(dataX[:train_size]), np.array(dataY[:train_size])
-	testX, testY = np.array(dataX[train_size:]), np.array(dataY[train_size:])
+	create_data(data) #TODO
 
 	## ======== Build model ======
 	with tf.Session() as sess:
@@ -159,8 +73,8 @@ if __name__ == '__main__':
 		saver = tf.train.Saver()
 		if not os.path.exists(DIRNAME + '/saved_graphs'):
 			os.makedirs(DIRNAME + '/saved_graphs')
-		saver.save(sess, DIRNAME + '/saved_graphs/' + args.model_name + '.chkp')
-		print("Saved")
+		path = saver.save(sess, DIRNAME + '/saved_graphs/' + args.model_name + ".ckpt")
+		print("Saved", path)
